Rename E2E to discovery
Lighthouse E2E still uses `dataplane` as tag/name for its E2E eventhough
we don't test dataplane at all. Rename it to avoid conflict with
submarine `dataplane` tests.
