+++ test/e2e/e2e_test.go
@@ -3,7 +3,7 @@ package e2e
 import (
 	"testing"
 
-	_ "github.com/submariner-io/lighthouse/test/e2e/dataplane"
+	_ "github.com/submariner-io/lighthouse/test/e2e/discovery"
 	"github.com/submariner-io/shipyard/test/e2e"
 )
